+++ src/utils/summoner.ts
@@ -5,6 +5,7 @@ import { BASTJInterpreterURI, BASTJInterpreter } from './interpreters-BAST'
 import { ProgramUnionURI, ProgramUnion } from './program'
 
 export interface M<E, A> extends Materialized<E, A, ProgramUnionURI, BASTJInterpreterURI> {}
+export interface UM<A> extends Materialized<unknown, A, ProgramUnionURI, BASTJInterpreterURI> {}
 
 export interface Prog<L, A> extends ProgramUnion<L, A> {}
 
@@ -12,7 +13,7 @@ interface Summons {
   summonAs: <L, A>(F: Prog<L, A>) => M<L, A>
   summonAsA: <A>() => <L>(F: Prog<L, A>) => M<L, A>
   summonAsL: <L>() => <A>(F: Prog<L, A>) => M<L, A>
-  summon: <A>(F: Prog<unknown, A>) => M<unknown, A>
+  summon: <A>(F: Prog<unknown, A>) => UM<A>
 }
 
 const { summonAs, summonAsA, summonAsL, summon } = makeSummoner(